.ipynb_checkpoints/requirements-checkpoint.txt
absl-py==0.12.0
aiohttp==3.7.4.
post0anyio==2.2.0
appdirs==1.4.4
Tensorflow-cpu==2.6.0
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.
1attrs==19.3.0
backcall==0.2.0
base58==2.1.0
beautifulsoup4==4.6.3
bleach==3.3.0
bs4==0.0.1
cachetools==4.2.1
catalogue==2.0.6
certifi==2019.11.28cffi==1.14.5
chardet==3.0.4
cmd2==1.5.0
configobj==5.0.6
configparser==5.0.2
decorator==4.4.2
dill==0.3.3
distro==1.4.0
ediblepickle==1.1.3
fake-useragent==0.1.11
google-auth==1.28.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
graphviz==0.16
grpcio==1.32.0
h5py==2.10.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
imageio==2.9.0
importlib-metadata==1.5.0
incremental==16.10.1

Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==18.0.1
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.0rc2
matplotlib==3.4.1
mistune==0.8.4
more-itertools==4.2.0
mpmath==1.2.1
multidict==5.1.0
natsort==7.1.1
nest-asyncio==1.5.1
netifaces==0.10.4
numpy==1.19.5
oauthlib==3.1.0
opteinsum==3.3.0
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parse==1.19.0
parso==0.8.2
pathy==0.6.0
patsy==0.5.1
pbr==5.5.1
pexpect==4.6.0
Pillow==8.2.0
plotly==4.14.3
preshed==3.0.5
prettytable==0.7.2
prometheus-client==0.10.0
prompt-toolkit==3.0.18
protobuf==3.15.7
ptyprocess==0.7.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.20
pydantic==1.8.2
pydeck==0.6.2
pyee==8.1.0
Pygments==2.8.1
PyHamcrest==1.9.0
PyJWT==1.7.1pymacaroons==0.13.0PyNaCl==1.3.0pyOpenSSL==19.0.0pyparsing==2.4.7
pyperclip==1.8.2pyppeteer==0.2.5
pyquery==1.4.3
pyrsistent==0.15.5pyserial==3.4python-dateutil==2.8.1python-dotenv==0.18.0
python-slugify==4.0.1
PythonQwt==0.9.0
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==22.0.3
QtPy==1.9.0
regex==2021.8.28
requests==2.22.0requests-futures==1.0.0
requests-html==0.10.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
retrying==1.3.3rsa==4.7.2
scikit-image==0.18.1scikit-learn==0.24.1scipy==1.6.2
seaborn==0.11.1SecretStorage==2.3.1selenium==3.141.0Send2Trash==1.5.0service-identity==18.1.0
simplejson==3.16.0six==1.15.0smart-open==5.2.1smmap==4.0.0sniffio==1.2.0spacy==3.1.2spacy-legacy==3.0.8srsly==2.4.1ssh-import-id==5.10statsmodels==0.12.2
stevedore==3.3.0
streamlit==0.83.0
systemd-python==234
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
text-unidecode==1.3
tfrecord==1.13
thinc==8.0.8
threadpoolctl==2.1.0
tifffile==2021.4.8
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
Twisted==18.9.0
typer==0.3.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.8
validators==0.18.2
vega-datasets==0.9.0
w3lib==1.22.0
wadllib==1.3.3
webencodings==0.5.1
websockets==8.1
Werkzeug==2.0.0
rc4widgetsnbextension==3.5.1
wrapt==1.12.1
yarl==1.6.3
zipp==1.0.0
zope.interface==4.7.1